+++ package.json
@@ -7,6 +7,7 @@
     "@skynexui/components": "^1.24.2",
     "@supabase/supabase-js": "^1.29.4",
     "next": "^12.0.9",
+    "polished": "^4.1.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
