+++ src/components/TitleWelcome/index.js
@@ -1,0 +1,19 @@
+// APENAS UM DEMONSTRATIVO DE COMO É POSSIVEL FAZER TAMBEM A COMPONENTIZAÇÃO
+
+// const Title = (props)=> {
+//   console.log('Prop', {props})
+//   const Tag = props.tag || h1;
+//   return (
+//     <>
+//       <Tag>{props.children}</Tag>
+//       <style jsx>{`
+//         ${Tag} {
+//           font-size:24px;
+//           font-weight:  500;
+//         }
+//         `}</style>
+//     </>
+//   )
+//   // ;
+// }
+// export default Title;
