+++ src/config.json
@@ -1,30 +1,30 @@
 {
-  "name": "Project - Discord (Ana)",
+  "name": "Chat Sharks",
   "theme": {
     "colors": {
       "primary": {
-        "050": "#E3F9E5",
-        "100": "#C1EAC5",
-        "200": "#A3D9A5",
-        "300": "#7BC47F",
-        "400": "#57AE5B",
-        "500": "#EA2B1F",
-        "600": "#2F8132",
-        "700": "#207227",
-        "800": "#0E5814",
-        "900": "#05400A"
+        "050": "#F8F8F8",
+        "100": "#ADE8F4",
+        "200": "#90E0EF",
+        "300": "#48CAE4",
+        "400": "#00B4D8",
+        "500": "#0096C7",
+        "600": "#0077B6",
+        "700": "#023E8A",
+        "800": "#03045E",
+        "900": "#13144E"
       },
       "neutrals": {
         "000": "#FFFFFF",
         "050": "#F5F7FA",
-        "100": "#E4E7EB",
-        "200": "#CBD2D9",
-        "300": "#9AA5B1",
-        "400": "#52667A",
-        "500": "#313D49",
-        "600": "#29333D",
-        "700": "#212931",
-        "800": "#181F25",
+        "100": "#E2E3FE",
+        "200": "#C6C7FD",
+        "300": "#A9ABFC",
+        "400": "#8D8EFB",
+        "500": "#202141",
+        "600": "#1D1E44",
+        "700": "#27273A",
+        "800": "#1D1E44",
         "900": "#101418",
         "999": "#080A0C"
       }
